Do you like my sword sword sword?
-Added Armory Class
-Added Weapon Class
-Made WorldObjects Objects
-General Cleanup
-Entity Events

+++ src/main/java/Tile.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 public class Tile {
     public static Map<Integer, List<String>> tileTags = new HashMap<>();
-    public static Map<Integer, Map<String, Integer>> tileValues = new HashMap<>();
+    public static Map<Integer, Map<String, Double>> tileValues = new HashMap<>();
     public static Map<Integer, Map<String, String>> tileTextures = new HashMap<>();
 
     public Tile(){
